+++ lib/stripe/object_types.rb
@@ -162,8 +162,8 @@ module Stripe
         V2::Billing::MeterEvent.object_name => V2::Billing::MeterEvent,
         V2::Billing::MeterEventAdjustment.object_name => V2::Billing::MeterEventAdjustment,
         V2::Billing::MeterEventSession.object_name => V2::Billing::MeterEventSession,
-        V2::Event.object_name => V2::Event,
-        V2::EventDestination.object_name => V2::EventDestination,
+        V2::Core::Event.object_name => V2::Core::Event,
+        V2::Core::EventDestination.object_name => V2::Core::EventDestination,
         # v2 object classes: The end of the section generated from our OpenAPI spec
       }
     end
