+++ lib/stripe/resources.rb
@@ -141,9 +141,9 @@ require "stripe/resources/treasury/transaction_entry"
 require "stripe/resources/v2/billing/meter_event"
 require "stripe/resources/v2/billing/meter_event_adjustment"
 require "stripe/resources/v2/billing/meter_event_session"
+require "stripe/resources/v2/core/event"
+require "stripe/resources/v2/core/event_destination"
 require "stripe/resources/v2/deleted_object"
-require "stripe/resources/v2/event"
-require "stripe/resources/v2/event_destination"
 require "stripe/resources/webhook_endpoint"
 require "stripe/events/v1_billing_meter_error_report_triggered_event"
 require "stripe/events/v1_billing_meter_no_meter_found_event"
