+++ stripe.gemspec
@@ -35,7 +35,7 @@ Gem::Specification.new do |s|
     # Handwritten RBIs
     # TODO(helenye): http://go/j/DEVSDK-2769
     %r{\Arbi/stripe/stripe_client.rbi\z},
-    %r{\Arbi/stripe/event_notification.rbi\z}
+    %r{\Arbi/stripe/resources/v2/core/event_notification.rbi\z}
   )
   s.files = `git ls-files`.split("\n").grep(included)
   s.bindir = "exe"
